requirements.txt
# Core dependencies
streamlit==1.31.0
python-dotenv==1.0.1
uuid==1.30

# Resume parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Database
psycopg2-binary==2.9.9

# Vector store
faiss-cpu==1.7.4
numpy==1.24.4
pandas==2.0.3

# API connections
groq==0.18.0  # For Groq API

# Other utilities
scikit-learn==1.3.2  # For potential additional vector operations

sentence_transformers